requirements.txt
shortuuid
pyyaml
loguru
uvicorn
fastapi
tortoise-orm[asyncmy]
redis-om
python-multipart
python-jose[cryptography]
pydantic[email]
passlib[bcrypt]